Extract createTextOptions module
