+++ src/renderPptx/units.ts
@@ -1,0 +1,5 @@
+export const PX_PER_IN = 96;
+
+export const pxToIn = (px: number) => px / PX_PER_IN;
+
+export const pxToPt = (px: number) => (px * 72) / PX_PER_IN;
